Add timing info to test output

+++ tests/produce_integration_test_data.py
@@ -9,6 +9,7 @@ fail at the tens-of-percent level.
 import glob
 import os
 import sys
+import tempfile
 
 import h5py
 from powerbox import get_power
@@ -17,6 +18,7 @@ from py21cmfast import AstroParams
 from py21cmfast import CosmoParams
 from py21cmfast import FlagOptions
 from py21cmfast import UserParams
+from py21cmfast import config
 from py21cmfast import global_params
 from py21cmfast import run_coeval
 from py21cmfast import run_lightcone
@@ -52,18 +54,28 @@ OPTIONS = (
     ],
     [8, {"N_THREADS": 2}],
     [10, {"PHOTON_CONS": True}],
-    [8, {"USE_MASS_DEPENDENT_ZETA": True, "PHOTON_CONS": True}],
     [
-        8,
+        8.5,
+        {
+            "USE_MASS_DEPENDENT_ZETA": True,
+            "PHOTON_CONS": True,
+            "z_heat_max": 25,
+            "zprime_step_factor": 1.1,
+        },
+    ],
+    [
+        9,
         {
             "USE_MASS_DEPENDENT_ZETA": True,
             "USE_TS_FLUCT": True,
             "INHOMO_RECO": True,
             "PHOTON_CONS": True,
+            "z_heat_max": 25,
+            "zprime_step_factor": 1.1,
         },
     ],
     [
-        8,
+        8.5,
         {
             "N_THREADS": 2,
             "USE_FFTW_WISDOM": True,
@@ -71,6 +83,8 @@ OPTIONS = (
             "INHOMO_RECO": True,
             "USE_TS_FLUCT": True,
             "PHOTON_CONS": True,
+            "z_heat_max": 25,
+            "zprime_step_factor": 1.1,
         },
     ],
 )
@@ -97,8 +111,6 @@ def get_all_options(redshift, **kwargs):
         "cosmo_params": cosmo_params,
         "astro_params": astro_params,
         "flag_options": flag_options,
-        "regenerate": True,
-        "write": False,
         "use_interp_perturb_field": kwargs.get("use_interp_perturb_field", False),
         "random_seed": SEED,
     }
@@ -120,7 +132,6 @@ def produce_coeval_power_spectra(redshift, **kwargs):
 
 def produce_lc_power_spectra(redshift, **kwargs):
     options = get_all_options(redshift, **kwargs)
-
     lightcone = run_lightcone(max_redshift=options["redshift"] + 2, **options)
 
     p, k = get_power(
@@ -139,15 +150,22 @@ def get_filename(redshift, **kwargs):
     return os.path.join(DATA_PATH, fname)
 
 
-def produce_power_spectra_for_tests(redshift, **kwargs):
-    k, p, coeval = produce_coeval_power_spectra(redshift, **kwargs)
-    k_l, p_l, lc = produce_lc_power_spectra(redshift, **kwargs)
-
+def produce_power_spectra_for_tests(redshift, force, direc, **kwargs):
     fname = get_filename(redshift, **kwargs)
 
     # Need to manually remove it, otherwise h5py tries to add to it
     if os.path.exists(fname):
-        os.remove(fname)
+        if force:
+            os.remove(fname)
+        else:
+            return fname
+
+    # For tests, we *don't* want to use cached boxes, but we also want to use the
+    # cache between the power spectra and lightcone. So we create a temporary
+    # directory in which to cache results.
+    with config.use(direc=direc):
+        k, p, coeval = produce_coeval_power_spectra(redshift, **kwargs)
+        k_l, p_l, lc = produce_lc_power_spectra(redshift, **kwargs)
 
     with h5py.File(fname, "w") as fl:
         for k, v in kwargs.items():
@@ -167,10 +185,48 @@ def produce_power_spectra_for_tests(redshift, **kwargs):
         fl["Tb"] = lc.global_brightness_temp
 
     print(f"Produced {fname} with {kwargs}")
+    return fname
 
 
 if __name__ == "__main__":
+    import logging
+
+    logger = logging.getLogger("21cmFAST")
+
+    lvl = "WARNING"
+    for arg in sys.argv:
+        if arg.startswith("--log"):
+            lvl = arg.split("--log")[-1]
+    lvl = getattr(logging, lvl)
+    logger.setLevel(lvl)
+
     global_params.ZPRIME_STEP_FACTOR = DEFAULT_ZPRIME_STEP_FACTOR
 
-    for redshift, kwargs in OPTIONS:
-        produce_power_spectra_for_tests(redshift, **kwargs)
+    force = "--force" in sys.argv
+    remove = "--no-clean" not in sys.argv
+
+    nums = range(len(OPTIONS))
+    for arg in sys.argv:
+        if arg.startswith("--nums="):
+            nums = [int(x) for x in arg.split("=")[-1].split(",")]
+            remove = False
+            force = True
+
+    # For tests, we *don't* want to use cached boxes, but we also want to use the
+    # cache between the power spectra and lightcone. So we create a temporary
+    # directory in which to cache results.
+    direc = tempfile.mkdtemp()
+    fnames = []
+    for redshift, kwargs in [OPTIONS[n] for n in nums]:
+        fnames.append(produce_power_spectra_for_tests(redshift, force, direc, **kwargs))
+
+    # Remove extra files that
+    if not nums:
+        all_files = glob.glob(os.path.join(DATA_PATH, "*"))
+        for fl in all_files:
+            if fl not in fnames:
+                if remove:
+                    print(f"Removing old file: {fl}")
+                    os.remove(fl)
+                else:
+                    print(f"File is now redundant and can be removed: {fl}")
